requirements.txt
flask==3.1.0
langchain-groq==0.2.1
python-dotenv==1.0.1
langchain==0.3.8
langchain_community==0.3.8
snowflake-sqlalchemy==1.7.0
snowflake-connector-python==3.12.3
pandas==2.2.3
flasgger==0.9.7.1